requirements.txt
boto3==1.11.5
botocore==1.14.5
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
docutils==0.15.2
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.16.0
idna==2.8
jmespath==0.9.4
KleenLogger==0.1.2
KleenTimer==0.1.2
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.1
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.7
